fix: Cropper typo
